Fix HAP task data race and WDT registration order
Route HomeKit TargetPosition and HoldPosition writes through
SomfyCommandQueue so the HAP library task no longer calls sendCommand()
directly — eliminating a data race with the main task.

Register the main task with the watchdog before somfy.begin() so that
esp_task_wdt_reset() calls during initialisation no longer fail with
"task not found".

Also update .gitignore to track docs/ instead of documentation/, and
add a cast to suppress a compiler warning in SomfyShade::clear().

+++ main/somfy/shade/SomfyShade.cpp
@@ -50,7 +50,7 @@ void SomfyShade::clear() {
   commandProcessor.stepSize = 100;
   movementTracker.lastMovement = 0;
   this->repeats = 1;
-  this->sortOrder = 255;
+  this->sortOrder = (int8_t)255;
 }
 
 bool SomfyShade::linkRemote(uint32_t address, uint16_t rollingCode) {
